+++ artifact_gen/lib/component/inatance.dart
@@ -2,6 +2,7 @@ import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:artifact_gen/builder.dart';
+import 'package:build/build.dart';
 
 class $ArtifactInstanceComponent implements $ArtifactBuilderOutput {
   const $ArtifactInstanceComponent();
@@ -12,6 +13,7 @@ class $ArtifactInstanceComponent implements $ArtifactBuilderOutput {
     ClassElement clazz,
     ConstructorElement ctor,
     List<FormalParameterElement> params,
+    BuildStep step,
   ) async {
     StringBuffer buf = StringBuffer();
     List<Uri> importUris = <Uri>[];