+++ example/lib/example.dart
@@ -1,36 +1,87 @@
 import 'package:artifact/artifact.dart';
+import 'package:example/gen/artifacts.gen.dart';
 
-// default is just `@artifact`
-@Artifact(compression: false, reflection: true)
-class Test {
-  final String aString;
-  final int aInt;
-  final AEnum anEnum;
-  final Sub aSub;
-  final List<Sub>? subList;
-  final Set<int> aSet;
+class Flag {
+  final String name;
 
-  @rename("m")
-  final Map<int, String> aMap;
-
-  const Test({
-    this.aString = 'default',
-    required this.aInt,
-    required this.anEnum,
-    required this.aSub,
-    required this.subList,
-    required this.aSet,
-    required this.aMap,
-  });
+  const Flag({required this.name});
 }
 
-@Artifact(compression: false, reflection: true)
-class Sub {
-  final int value;
+const Flag basicFlag = Flag(name: "Basic Flag 42");
 
-  Sub({this.value = 42});
+class Start {
+  const Start();
 }
 
-enum AEnum { first, second, third }
+mixin AMixin {}
 
-void main() {}
+abstract class AInterface {}
+
+const Artifact art = Artifact(compression: true);
+
+@art
+@Flag(name: "This is A")
+class A with AMixin implements AInterface {
+  @basicFlag
+  final int a;
+
+  A({this.a = 42});
+
+  @basicFlag
+  @Start()
+  void foo() {
+    print("A.foo called!");
+  }
+
+  void bar() {}
+}
+
+@art
+class B {
+  int? mutableField;
+
+  final int a;
+
+  B({this.a = 42});
+}
+
+@basicFlag
+@art
+class C {
+  final int a;
+
+  C({this.a = 42});
+}
+
+@art
+@Flag(name: "This is D")
+@basicFlag
+class D {
+  final int a;
+
+  D({this.a = 42});
+
+  @Start()
+  void ass() {
+    print("D.ass called!");
+  }
+
+  @Start()
+  void bad() {
+    print("D.bad called!");
+  }
+}
+
+void main() {
+  // 1. Loop over all classes with @Flag annotation
+  for ($AClass clazz in $artifactMirror.withAnnotation<Flag>().values) {
+    // Create a new instance of this class
+    Object instance = clazz.construct();
+
+    // Find all methods in this class with the @Start annotation
+    for ($AMth method in clazz.annotatedMethods<Start>()) {
+      // Invoke the method with no parameters using the instance we just created
+      method(instance, MethodParameters());
+    }
+  }
+}